requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
distlib==0.3.8
Django==4.2.14
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
filelock==3.15.4
idna==3.7
inflection==0.5.1
Markdown==3.6
packaging==24.1
pillow==10.4.0
pipenv==2024.0.1
platformdirs==4.2.2
PyJWT==2.9.0
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
setuptools==72.1.0
sqlparse==0.5.1
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
virtualenv==20.26.3
